+++ applications/ShapeOptimizationApplication/python_scripts/algorithms/algorithm_steepest_descent.py
@@ -15,29 +15,31 @@ from __future__ import print_function, absolute_import, division
 # Kratos Core and Apps
 import KratosMultiphysics as KM
 import KratosMultiphysics.ShapeOptimizationApplication as KSO
-from KratosMultiphysics.LinearSolversApplication import dense_linear_solver_factory
 
 # Additional imports
-from .algorithm_base import OptimizationAlgorithm
-from . import mapper_factory
-from . import data_logger_factory
-from .custom_timer import Timer
-from .custom_variable_utilities import WriteDictionaryDataOnNodalVariable
+from KratosMultiphysics.ShapeOptimizationApplication.algorithms.algorithm_base import OptimizationAlgorithm
+from KratosMultiphysics.ShapeOptimizationApplication import mapper_factory
+from KratosMultiphysics.ShapeOptimizationApplication.loggers import data_logger_factory
+from KratosMultiphysics.ShapeOptimizationApplication.custom_timer import Timer
+from KratosMultiphysics.ShapeOptimizationApplication.custom_variable_utilities import WriteDictionaryDataOnNodalVariable
 
 # ==============================================================================
-class AlgorithmGradientProjection(OptimizationAlgorithm):
+class AlgorithmSteepestDescent(OptimizationAlgorithm):
     # --------------------------------------------------------------------------
     def __init__(self, optimization_settings, analyzer, communicator, model_part_controller):
         default_algorithm_settings = KM.Parameters("""
         {
-            "name"                    : "penalized_projection",
-            "max_correction_share"    : 0.75,
-            "max_iterations"          : 100,
-            "relative_tolerance"      : 1e-3,
+            "name"               : "steepest_descent",
+            "max_iterations"     : 100,
+            "relative_tolerance" : 1e-3,
+            "gradient_tolerance" : 1e-5,
             "line_search" : {
                 "line_search_type"           : "manual_stepping",
                 "normalize_search_direction" : true,
-                "step_size"                  : 1.0
+                "step_size"                  : 1.0,
+                "estimation_tolerance"       : 0.1,
+                "increase_factor"            : 1.1,
+                "max_increase_factor"        : 10.0
             }
         }""")
         self.algorithm_settings =  optimization_settings["optimization_algorithm"]
@@ -57,30 +59,27 @@ class AlgorithmGradientProjection(OptimizationAlgorithm):
 
         self.objectives = optimization_settings["objectives"]
         self.constraints = optimization_settings["constraints"]
-        self.constraint_gradient_variables = {}
-        for itr, constraint in enumerate(self.constraints):
-            self.constraint_gradient_variables.update({
-                constraint["identifier"].GetString() : {
-                    "gradient": KM.KratosGlobals.GetVariable("DC"+str(itr+1)+"DX"),
-                    "mapped_gradient": KM.KratosGlobals.GetVariable("DC"+str(itr+1)+"DX_MAPPED")
-                }
-            })
-        self.max_correction_share = self.algorithm_settings["max_correction_share"].GetDouble()
 
-        self.step_size = self.algorithm_settings["line_search"]["step_size"].GetDouble()
+        self.previos_objective_value = None
+
         self.max_iterations = self.algorithm_settings["max_iterations"].GetInt() + 1
         self.relative_tolerance = self.algorithm_settings["relative_tolerance"].GetDouble()
+        self.gradient_tolerance = self.algorithm_settings["gradient_tolerance"].GetDouble()
+        self.line_search_type = self.algorithm_settings["line_search"]["line_search_type"].GetString()
+        self.estimation_tolerance = self.algorithm_settings["line_search"]["estimation_tolerance"].GetDouble()
+        self.step_size = self.algorithm_settings["line_search"]["step_size"].GetDouble()
+        self.increase_factor = self.algorithm_settings["line_search"]["increase_factor"].GetDouble()
+        self.max_step_size = self.step_size*self.algorithm_settings["line_search"]["max_increase_factor"].GetDouble()
 
         self.optimization_model_part = model_part_controller.GetOptimizationModelPart()
         self.optimization_model_part.AddNodalSolutionStepVariable(KSO.SEARCH_DIRECTION)
-        self.optimization_model_part.AddNodalSolutionStepVariable(KSO.CORRECTION)
 
     # --------------------------------------------------------------------------
     def CheckApplicability(self):
         if self.objectives.size() > 1:
-            raise RuntimeError("Gradient projection algorithm only supports one objective function!")
-        if self.constraints.size() == 0:
-            raise RuntimeError("Gradient projection algorithm requires definition of at least one constraint!")
+            raise RuntimeError("Steepest descent algorithm only supports one objective function!")
+        if self.constraints.size() > 0:
+            raise RuntimeError("Steepest descent algorithm does not allow for any constraints!")
 
     # --------------------------------------------------------------------------
     def InitializeOptimizationLoop(self):
@@ -106,7 +105,7 @@ class AlgorithmGradientProjection(OptimizationAlgorithm):
         for self.optimization_iteration in range(1,self.max_iterations):
             KM.Logger.Print("")
             KM.Logger.Print("===============================================================================")
-            KM.Logger.PrintInfo("ShapeOpt", timer.GetTimeStamp(), ": Starting optimization iteration ", self.optimization_iteration)
+            KM.Logger.PrintInfo("ShapeOpt", "",timer.GetTimeStamp(), ": Starting optimization iteration ",self.optimization_iteration)
             KM.Logger.Print("===============================================================================\n")
 
             timer.StartNewLap()
@@ -114,6 +113,9 @@ class AlgorithmGradientProjection(OptimizationAlgorithm):
             self.__initializeNewShape()
 
             self.__analyzeShape()
+
+            if self.line_search_type == "adaptive_stepping" and self.optimization_iteration > 1:
+                self.__adjustStepSize()
 
             self.__computeShapeUpdate()
 
@@ -145,145 +147,75 @@ class AlgorithmGradientProjection(OptimizationAlgorithm):
         self.communicator.requestValueOf(self.objectives[0]["identifier"].GetString())
         self.communicator.requestGradientOf(self.objectives[0]["identifier"].GetString())
 
-        for constraint in self.constraints:
-            con_id =  constraint["identifier"].GetString()
-            self.communicator.requestValueOf(con_id)
-            self.communicator.requestGradientOf(con_id)
-
         self.analyzer.AnalyzeDesignAndReportToCommunicator(self.optimization_model_part, self.optimization_iteration, self.communicator)
 
-        # compute normals only if required
-        surface_normals_required = self.objectives[0]["project_gradient_on_surface_normals"].GetBool()
-        for constraint in self.constraints:
-            if constraint["project_gradient_on_surface_normals"].GetBool():
-                surface_normals_required = True
-
-        if surface_normals_required:
-            self.model_part_controller.ComputeUnitSurfaceNormals()
-
-        # project and damp objective gradients
         objGradientDict = self.communicator.getStandardizedGradient(self.objectives[0]["identifier"].GetString())
         WriteDictionaryDataOnNodalVariable(objGradientDict, self.optimization_model_part, KSO.DF1DX)
 
         if self.objectives[0]["project_gradient_on_surface_normals"].GetBool():
+            self.model_part_controller.ComputeUnitSurfaceNormals()
             self.model_part_controller.ProjectNodalVariableOnUnitSurfaceNormals(KSO.DF1DX)
 
         self.model_part_controller.DampNodalVariableIfSpecified(KSO.DF1DX)
 
-        # project and damp constraint gradients
-        for constraint in self.constraints:
-            con_id = constraint["identifier"].GetString()
-            conGradientDict = self.communicator.getStandardizedGradient(con_id)
-            gradient_variable = self.constraint_gradient_variables[con_id]["gradient"]
-            WriteDictionaryDataOnNodalVariable(conGradientDict, self.optimization_model_part, gradient_variable)
+    # --------------------------------------------------------------------------
+    def __adjustStepSize(self):
+        current_a = self.step_size
 
-            if constraint["project_gradient_on_surface_normals"].GetBool():
-                self.model_part_controller.ProjectNodalVariableOnUnitSurfaceNormals(gradient_variable)
+        # Compare actual and estimated improvement using linear information from the previos step
+        dfda1 = 0.0
+        for node in self.design_surface.Nodes:
+            # The following variables are not yet updated and therefore contain the information from the previos step
+            s1 = node.GetSolutionStepValue(KSO.SEARCH_DIRECTION)
+            dfds1 = node.GetSolutionStepValue(KSO.DF1DX_MAPPED)
+            dfda1 += s1[0]*dfds1[0] + s1[1]*dfds1[1] + s1[2]*dfds1[2]
 
-            self.model_part_controller.DampNodalVariableIfSpecified(gradient_variable)
+        f2 = self.communicator.getStandardizedValue(self.objectives[0]["identifier"].GetString())
+        f1 = self.previos_objective_value
+
+        df_actual = f2 - f1
+        df_estimated = current_a*dfda1
+
+        # Adjust step size if necessary
+        if f2 < f1:
+            estimation_error = (df_actual-df_estimated)/df_actual
+
+            # Increase step size if estimation based on linear extrapolation matches the actual improvement within a specified tolerance
+            if estimation_error < self.estimation_tolerance:
+                new_a = min(current_a*self.increase_factor, self.max_step_size)
+
+            # Leave step size unchanged if a nonliner change in the objective is observed but still a descent direction is obtained
+            else:
+                new_a = current_a
+        else:
+            # Search approximation of optimal step using interpolation
+            a = current_a
+            corrected_step_size = - 0.5 * dfda1 * a**2 / (f2 - f1 - dfda1 * a )
+
+            # Starting from the new design, and assuming an opposite gradient direction, the step size to the approximated optimum behaves reciprocal
+            new_a = current_a-corrected_step_size
+
+        self.step_size = new_a
 
     # --------------------------------------------------------------------------
     def __computeShapeUpdate(self):
         self.mapper.Update()
         self.mapper.InverseMap(KSO.DF1DX, KSO.DF1DX_MAPPED)
 
-        for constraint in self.constraints:
-            con_id = constraint["identifier"].GetString()
-            gradient_variable = self.constraint_gradient_variables[con_id]["gradient"]
-            mapped_gradient_variable = self.constraint_gradient_variables[con_id]["mapped_gradient"]
-            self.mapper.InverseMap(gradient_variable, mapped_gradient_variable)
-
-        self.__computeControlPointUpdate()
+        self.optimization_utilities.ComputeSearchDirectionSteepestDescent()
+        self.optimization_utilities.ComputeControlPointUpdate(self.step_size)
 
         self.mapper.Map(KSO.CONTROL_POINT_UPDATE, KSO.SHAPE_UPDATE)
         self.model_part_controller.DampNodalVariableIfSpecified(KSO.SHAPE_UPDATE)
 
     # --------------------------------------------------------------------------
-    def __computeControlPointUpdate(self):
-        """adapted from https://msulaiman.org/onewebmedia/GradProj_2.pdf"""
-        gp_utilities = self.optimization_utilities
-
-        g_a, g_a_variables = self.__getActiveConstraints()
-
-        KM.Logger.PrintInfo("ShapeOpt", "Assemble vector of objective gradient.")
-        nabla_f = KM.Vector()
-        s = KM.Vector()
-        gp_utilities.AssembleVector(nabla_f, KSO.DF1DX_MAPPED)
-
-        if len(g_a) == 0:
-            KM.Logger.PrintInfo("ShapeOpt", "No constraints active, use negative objective gradient as search direction.")
-            s = nabla_f * (-1.0)
-            s *= self.step_size / s.norm_inf()
-            gp_utilities.AssignVectorToVariable(s, KSO.SEARCH_DIRECTION)
-            gp_utilities.AssignVectorToVariable([0.0]*len(s), KSO.CORRECTION)
-            gp_utilities.AssignVectorToVariable(s, KSO.CONTROL_POINT_UPDATE)
-            return
-
-
-        KM.Logger.PrintInfo("ShapeOpt", "Assemble matrix of constraint gradient.")
-        N = KM.Matrix()
-        gp_utilities.AssembleMatrix(N, g_a_variables)  # TODO check if gradients are 0.0! - in cpp
-
-        settings = KM.Parameters('{ "solver_type" : "LinearSolversApplication.dense_col_piv_householder_qr" }')
-        solver = dense_linear_solver_factory.ConstructSolver(settings)
-
-        KM.Logger.PrintInfo("ShapeOpt", "Calculate projected search direction and correction.")
-        c = KM.Vector()
-        gp_utilities.CalculateProjectedSearchDirectionAndCorrection(
-            nabla_f,
-            N,
-            g_a,
-            solver,
-            s,
-            c)
-
-        if c.norm_inf() != 0.0:
-            if c.norm_inf() <= self.max_correction_share * self.step_size:
-                delta = self.step_size - c.norm_inf()
-                s *= delta/s.norm_inf()
-            else:
-                KM.Logger.PrintWarning("ShapeOpt", f"Correction is scaled down from {c.norm_inf()} to {self.max_correction_share * self.step_size}.")
-                c *= self.max_correction_share * self.step_size / c.norm_inf()
-                s *= (1.0 - self.max_correction_share) * self.step_size / s.norm_inf()
-        else:
-            s *= self.step_size / s.norm_inf()
-
-        gp_utilities.AssignVectorToVariable(s, KSO.SEARCH_DIRECTION)
-        gp_utilities.AssignVectorToVariable(c, KSO.CORRECTION)
-        gp_utilities.AssignVectorToVariable(s+c, KSO.CONTROL_POINT_UPDATE)
-
-    # --------------------------------------------------------------------------
-    def __getActiveConstraints(self):
-        active_constraint_values = []
-        active_constraint_variables = []
-
-        for constraint in self.constraints:
-            if self.__isConstraintActive(constraint):
-                identifier = constraint["identifier"].GetString()
-                constraint_value = self.communicator.getStandardizedValue(identifier)
-                active_constraint_values.append(constraint_value)
-                active_constraint_variables.append(
-                    self.constraint_gradient_variables[identifier]["mapped_gradient"])
-
-        return active_constraint_values, active_constraint_variables
-
-    # --------------------------------------------------------------------------
-    def __isConstraintActive(self, constraint):
-        identifier = constraint["identifier"].GetString()
-        constraint_value = self.communicator.getStandardizedValue(identifier)
-        if constraint["type"].GetString() == "=":
-            return True
-        elif constraint_value >= 0:
-            return True
-        else:
-            return False
-
-    # --------------------------------------------------------------------------
     def __logCurrentOptimizationStep(self):
+        self.previos_objective_value = self.communicator.getStandardizedValue(self.objectives[0]["identifier"].GetString())
+        self.norm_objective_gradient = self.optimization_utilities.ComputeL2NormOfNodalVariable(KSO.DF1DX_MAPPED)
+
         additional_values_to_log = {}
         additional_values_to_log["step_size"] = self.step_size
-        additional_values_to_log["inf_norm_s"] = self.optimization_utilities.ComputeMaxNormOfNodalVariable(KSO.SEARCH_DIRECTION)
-        additional_values_to_log["inf_norm_c"] = self.optimization_utilities.ComputeMaxNormOfNodalVariable(KSO.CORRECTION)
+        additional_values_to_log["norm_objective_gradient"] = self.norm_objective_gradient
         self.data_logger.LogCurrentValues(self.optimization_iteration, additional_values_to_log)
         self.data_logger.LogCurrentDesign(self.optimization_iteration)
 
@@ -291,12 +223,20 @@ class AlgorithmGradientProjection(OptimizationAlgorithm):
     def __isAlgorithmConverged(self):
 
         if self.optimization_iteration > 1 :
-
             # Check if maximum iterations were reached
             if self.optimization_iteration == self.max_iterations:
                 KM.Logger.Print("")
                 KM.Logger.PrintInfo("ShapeOpt", "Maximal iterations of optimization problem reached!")
                 return True
+
+            # Check gradient norm
+            if self.optimization_iteration == 2:
+                self.initial_norm_objective_gradient = self.norm_objective_gradient
+            else:
+                if self.norm_objective_gradient < self.gradient_tolerance*self.initial_norm_objective_gradient:
+                    KM.Logger.Print("")
+                    KM.Logger.PrintInfo("ShapeOpt", "Optimization problem converged as gradient norm reached specified tolerance of ",self.gradient_tolerance)
+                    return True
 
             # Check for relative tolerance
             relative_change_of_objective_value = self.data_logger.GetValues("rel_change_objective")[self.optimization_iteration]
