+++ applications/ShapeOptimizationApplication/python_scripts/loggers/value_logger_gradient_projection.py
@@ -4,8 +4,7 @@
 #  License:         BSD License
 #                   license: ShapeOptimizationApplication/license.txt
 #
-#  Main authors:    Baumgaertner Daniel, https://github.com/dbaumgaertner
-#                   Geiser Armin, https://github.com/armingeiser
+#  Main authors:    Geiser Armin, https://github.com/armingeiser
 #
 # ==============================================================================
 
@@ -13,33 +12,32 @@
 import KratosMultiphysics as KM
 
 # Import logger base classes
-from .value_logger_base import ValueLogger
+from KratosMultiphysics.ShapeOptimizationApplication.loggers.value_logger_base import ValueLogger
 
 # Import additional libraries
 import csv
-from .custom_timer import Timer
+from KratosMultiphysics.ShapeOptimizationApplication.custom_timer import Timer
 
 # ==============================================================================
-class ValueLoggerBeadOptimization( ValueLogger ):
+class ValueLoggerGradientProjection( ValueLogger ):
     # --------------------------------------------------------------------------
     def InitializeLogging( self ):
         with open(self.complete_log_file_name, 'w') as csvfile:
             historyWriter = csv.writer(csvfile, delimiter=',',quotechar='|',quoting=csv.QUOTE_MINIMAL)
             row = []
-            row.append("{:>12s}".format("overall_itr"))
-            row.append("{:>10s}".format("outer_itr"))
-            row.append("{:>10s}".format("inner_itr"))
-            row.append("{:>13s}".format("l"))
-            row.append("{:>13s}".format("dl_rel[%]"))
+            row.append("{:>4s}".format("itr"))
             row.append("{:>13s}".format("f"))
             row.append("{:>13s}".format("df_abs[%]"))
             row.append("{:>13s}".format("df_rel[%]"))
-            row.append("{:>14s}".format("max_norm_dfda"))
-            row.append("{:>13s}".format("lambda_p"))
-            row.append("{:>13s}".format("p"))
-            row.append("{:>13s}".format("p_scaling"))
-            row.append("{:>13s}".format("p_fac"))
+
+            for itr in range(self.constraints.size()):
+                con_type = self.constraints[itr]["type"].GetString()
+                row.append("{:>13s}".format("c"+str(itr+1)+": "+con_type))
+                row.append("{:>13s}".format("c"+str(itr+1)+"_ref"))
+
             row.append("{:>13s}".format("step_size"))
+            row.append("{:>13s}".format("inf_norm_s"))
+            row.append("{:>13s}".format("inf_norm_c"))
             row.append("{:>25s}".format("time_stamp"))
             historyWriter.writerow(row)
 
@@ -47,32 +45,35 @@ class ValueLoggerBeadOptimization( ValueLogger ):
     def _WriteCurrentValuesToConsole( self ):
         objective_id = self.objectives[0]["identifier"].GetString()
         KM.Logger.Print("")
-        KM.Logger.PrintInfo("ShapeOpt", "Current value of objective = ","{:> .5E}".format(self.history["response_value"][objective_id][self.current_index]))
+        KM.Logger.PrintInfo("ShapeOpt", "Current value of objective = ", "{:> .5E}".format(self.history["response_value"][objective_id][self.current_index]))
 
-        KM.Logger.PrintInfo("ShapeOpt", "Absolut change of objective = ","{:> .5E}".format(self.history["abs_change_objective"][self.current_index])," [%]")
+        KM.Logger.PrintInfo("ShapeOpt", "Absolute change of objective = ","{:> .5E}".format(self.history["abs_change_objective"][self.current_index])," [%]")
         KM.Logger.PrintInfo("ShapeOpt", "Relative change of objective = ","{:> .5E}".format(self.history["rel_change_objective"][self.current_index])," [%]\n")
-        KM.Logger.PrintInfo("ShapeOpt", "Value of penalty term = ","{:> .5E}".format(self.history["penalty_value"][self.current_index]),"\n")
+
+        for itr in range(self.constraints.size()):
+            constraint_id = self.constraints[itr]["identifier"].GetString()
+            KM.Logger.PrintInfo("ShapeOpt", "Value of C"+str(itr+1)+" = ", "{:> .5E}".format(self.history["response_value"][constraint_id][self.current_index]))
 
     # --------------------------------------------------------------------------
     def _WriteCurrentValuesToFile( self ):
         with open(self.complete_log_file_name, 'a') as csvfile:
             historyWriter = csv.writer(csvfile, delimiter=',',quotechar='|',quoting=csv.QUOTE_MINIMAL)
             row = []
-            row.append("{:>12d}".format(self.current_index))
-            row.append("{:>10d}".format(self.history["outer_iteration"][self.current_index]))
-            row.append("{:>10d}".format(self.history["inner_iteration"][self.current_index]))
+            row.append("{:>4d}".format(self.current_index))
+
             objective_id = self.objectives[0]["identifier"].GetString()
-            row.append(" {:> .5E}".format(self.history["lagrange_value"][self.current_index]))
-            row.append(" {:> .5E}".format(self.history["lagrange_value_relative_change"][self.current_index]))
             row.append(" {:> .5E}".format(self.history["response_value"][objective_id][self.current_index]))
             row.append(" {:> .5E}".format(self.history["abs_change_objective"][self.current_index]))
             row.append(" {:> .5E}".format(self.history["rel_change_objective"][self.current_index]))
-            row.append("  {:> .5E}".format(self.history["max_norm_objective_gradient"][self.current_index]))
-            row.append(" {:> .5E}".format(self.history["penalty_lambda"][self.current_index]))
-            row.append(" {:> .5E}".format(self.history["penalty_value"][self.current_index]))
-            row.append(" {:> .5E}".format(self.history["penalty_scaling"][self.current_index]))
-            row.append(" {:> .5E}".format(self.history["penalty_factor"][self.current_index]))
+
+            for itr in range(self.constraints.size()):
+                constraint_id = self.constraints[itr]["identifier"].GetString()
+                row.append(" {:> .5E}".format(self.history["response_value"][constraint_id][self.current_index]))
+                row.append(" {:> .5E}".format(self.communicator.getReferenceValue(constraint_id)))
+
             row.append(" {:> .5E}".format(self.history["step_size"][self.current_index]))
+            row.append(" {:> .5E}".format(self.history["inf_norm_s"][self.current_index]))
+            row.append(" {:> .5E}".format(self.history["inf_norm_c"][self.current_index]))
             row.append("{:>25}".format(Timer().GetTimeStamp()))
             historyWriter.writerow(row)
