+++ applications/ShapeOptimizationApplication/python_scripts/loggers/value_logger_steepest_descent.py
@@ -4,7 +4,8 @@
 #  License:         BSD License
 #                   license: ShapeOptimizationApplication/license.txt
 #
-#  Main authors:    Geiser Armin, https://github.com/armingeiser
+#  Main authors:    Baumgaertner Daniel, https://github.com/dbaumgaertner
+#                   Geiser Armin, https://github.com/armingeiser
 #
 # ==============================================================================
 
@@ -12,14 +13,14 @@
 import KratosMultiphysics as KM
 
 # Import logger base classes
-from .value_logger_base import ValueLogger
+from KratosMultiphysics.ShapeOptimizationApplication.loggers.value_logger_base import ValueLogger
 
 # Import additional libraries
 import csv
-from .custom_timer import Timer
+from KratosMultiphysics.ShapeOptimizationApplication.custom_timer import Timer
 
 # ==============================================================================
-class ValueLoggerGradientProjection( ValueLogger ):
+class ValueLoggerSteepestDescent( ValueLogger ):
     # --------------------------------------------------------------------------
     def InitializeLogging( self ):
         with open(self.complete_log_file_name, 'w') as csvfile:
@@ -29,15 +30,8 @@ class ValueLoggerGradientProjection( ValueLogger ):
             row.append("{:>13s}".format("f"))
             row.append("{:>13s}".format("df_abs[%]"))
             row.append("{:>13s}".format("df_rel[%]"))
-
-            for itr in range(self.constraints.size()):
-                con_type = self.constraints[itr]["type"].GetString()
-                row.append("{:>13s}".format("c"+str(itr+1)+": "+con_type))
-                row.append("{:>13s}".format("c"+str(itr+1)+"_ref"))
-
+            row.append("{:>13s}".format("norm_df"))
             row.append("{:>13s}".format("step_size"))
-            row.append("{:>13s}".format("inf_norm_s"))
-            row.append("{:>13s}".format("inf_norm_c"))
             row.append("{:>25s}".format("time_stamp"))
             historyWriter.writerow(row)
 
@@ -47,12 +41,8 @@ class ValueLoggerGradientProjection( ValueLogger ):
         KM.Logger.Print("")
         KM.Logger.PrintInfo("ShapeOpt", "Current value of objective = ", "{:> .5E}".format(self.history["response_value"][objective_id][self.current_index]))
 
-        KM.Logger.PrintInfo("ShapeOpt", "Absolute change of objective = ","{:> .5E}".format(self.history["abs_change_objective"][self.current_index])," [%]")
+        KM.Logger.PrintInfo("ShapeOpt", "Absolut change of objective = ","{:> .5E}".format(self.history["abs_change_objective"][self.current_index])," [%]")
         KM.Logger.PrintInfo("ShapeOpt", "Relative change of objective = ","{:> .5E}".format(self.history["rel_change_objective"][self.current_index])," [%]\n")
-
-        for itr in range(self.constraints.size()):
-            constraint_id = self.constraints[itr]["identifier"].GetString()
-            KM.Logger.PrintInfo("ShapeOpt", "Value of C"+str(itr+1)+" = ", "{:> .5E}".format(self.history["response_value"][constraint_id][self.current_index]))
 
     # --------------------------------------------------------------------------
     def _WriteCurrentValuesToFile( self ):
@@ -66,14 +56,8 @@ class ValueLoggerGradientProjection( ValueLogger ):
             row.append(" {:> .5E}".format(self.history["abs_change_objective"][self.current_index]))
             row.append(" {:> .5E}".format(self.history["rel_change_objective"][self.current_index]))
 
-            for itr in range(self.constraints.size()):
-                constraint_id = self.constraints[itr]["identifier"].GetString()
-                row.append(" {:> .5E}".format(self.history["response_value"][constraint_id][self.current_index]))
-                row.append(" {:> .5E}".format(self.communicator.getReferenceValue(constraint_id)))
-
+            row.append(" {:> .5E}".format(self.history["norm_objective_gradient"][self.current_index]))
             row.append(" {:> .5E}".format(self.history["step_size"][self.current_index]))
-            row.append(" {:> .5E}".format(self.history["inf_norm_s"][self.current_index]))
-            row.append(" {:> .5E}".format(self.history["inf_norm_c"][self.current_index]))
             row.append("{:>25}".format(Timer().GetTimeStamp()))
             historyWriter.writerow(row)
