+++ applications/ShapeOptimizationApplication/tests/algorithm_gradient_projection_test/run_test.py
@@ -2,7 +2,7 @@
 import KratosMultiphysics as KM
 
 # Additional imports
-from KratosMultiphysics.ShapeOptimizationApplication.analyzer_base import AnalyzerBaseClass
+from KratosMultiphysics.ShapeOptimizationApplication.analyzers.analyzer_base import AnalyzerBaseClass
 from KratosMultiphysics.ShapeOptimizationApplication import optimizer_factory
 from KratosMultiphysics.KratosUnittest import TestCase
 import KratosMultiphysics.kratos_utilities as kratos_utilities
