+++ applications/ShapeOptimizationApplication/tests/opt_process_vertex_morphing_test/tent_analyzer.py
@@ -13,7 +13,7 @@
 #  _______________/        \_______________
 #  |----- 15 -----|-- 10 --|----- 15 -----|
 
-from KratosMultiphysics.ShapeOptimizationApplication.analyzer_base import AnalyzerBaseClass
+from KratosMultiphysics.ShapeOptimizationApplication.analyzers.analyzer_base import AnalyzerBaseClass
 
 class CustomAnalyzer(AnalyzerBaseClass):
