+++ applications/ShapeOptimizationApplication/tests/remeshing_opt_process_test/run_test.py
@@ -3,14 +3,14 @@ import KratosMultiphysics as KM
 from KratosMultiphysics.compare_two_files_check_process import CompareTwoFilesCheckProcess
 
 # Additional imports
-from KratosMultiphysics.ShapeOptimizationApplication.analyzer_base import AnalyzerBaseClass
+from KratosMultiphysics.ShapeOptimizationApplication.analyzers.analyzer_base import AnalyzerBaseClass
 from KratosMultiphysics.ShapeOptimizationApplication import optimizer_factory
 from KratosMultiphysics.KratosUnittest import TestCase
 import KratosMultiphysics.kratos_utilities as kratos_utilities
 import csv, os
 
 try:
-    from .custom_analyzer_y import CustomAnalyzer
+    from KratosMultiphysics.ShapeOptimizationApplication.analyzers.custom_analyzer_y import CustomAnalyzer
 except ImportError:
     from custom_analyzer_y import CustomAnalyzer
