+++ applications/ShapeOptimizationApplication/tests/trust_region_projector_test/run_test.py
@@ -3,7 +3,7 @@ import KratosMultiphysics as KM
 
 # Additional imports
 from KratosMultiphysics.KratosUnittest import TestCase
-from KratosMultiphysics.ShapeOptimizationApplication.algorithm_trust_region import Projector
+from KratosMultiphysics.ShapeOptimizationApplication.algorithms.algorithm_trust_region import Projector
 from KratosMultiphysics.ShapeOptimizationApplication.custom_math import PerformBisectioning
 import csv
